+++ NEWS.md
@@ -2,6 +2,9 @@
 
 - Fixed a bug where `hb_transform()` would give an error with missing values.
 
+- Added a new function `outliers()` to replace existing outlier functions, which
+are now deprecated.
+
 ## gpindex 0.6.3
 
 - Bumped minimum version of R to >= 4.1.